+++ interfaces/screen/inventory_menu.h
@@ -2,28 +2,7 @@
 #include "menu_button.h"
 #include "screen.h"
 #include "inventory.h"
-
-class GraphicSlot {
-public:
-    std::shared_ptr<Slot> slot;
-    sf::Text gr_amount;
-    sf::Sprite slot_sprite;
-
-    bool gr_amount_offset = false;
-    void show_slot(sf::RenderWindow& window);
-};
-
-class GraphicInventory {
-public:
-    std::vector<std::shared_ptr<GraphicSlot>> gr_items_array;
-    sf::Vector2f pos;
-    void build_inventory(const std::vector<std::shared_ptr<Slot>>& items, float x, float y);
-    std::shared_ptr<GraphicSlot>& operator[](std::size_t index);
-    void set_pos(float x, float y);
-    std::size_t size() const {
-        return gr_items_array.size();
-    }
-};
+#include "graphic_inventory.h"
 
 class InventoryMenu : public Screen {
 public:
@@ -34,8 +13,8 @@ public:
 
     MenuButton b_exit;
     std::vector<MenuButton*> buttons;
-    static inline GraphicInventory gr_inventory;
-    static inline sf::Text gr_money;
+    static inline std::shared_ptr<GraphicInventory> gr_inventory;
+    static inline std::shared_ptr<sf::Text> gr_money;
 
     virtual View_mode Run(sf::RenderWindow& window) override;
     static void update_graphic_inventory(const std::vector<std::shared_ptr<Slot>>& items_array, int _money);
