+++ interfaces/screen/trade_menu.cpp
@@ -49,31 +49,34 @@ View_mode TradeMenu::Run(sf::RenderWindow& window) {
     }
 }
 
-void TradeMenu::bind(const GraphicInventory& _gr_inventory_player,
-                     const sf::Text& _gr_money_player) {
-    gr_inventory_player = _gr_inventory_player;
-    gr_money_player = _gr_money_player;
+void TradeMenu::bind(std::shared_ptr<GraphicInventory> _gr_inventory_player,
+                     std::shared_ptr<sf::Text> _gr_money_player) {
+
+    gr_inventory_player = std::shared_ptr<GraphicInventory>(_gr_inventory_player);
+    gr_money_player = std::shared_ptr<sf::Text>(_gr_money_player);
 }
 
 void TradeMenu::update_graphic_inventories(const std::vector<std::shared_ptr<Slot>>& items_array_player,
                                          const std::vector<std::shared_ptr<Slot>>& items_array_trader,
                                          int _money_player, int _money_trader) {
 
-    gr_inventory_player.set_pos(300.f, 200.f);
+    gr_inventory_player->set_pos(300.f, 200.f);
+    gr_money_player->setPosition({300.f, 130.f});
 
     for (std::size_t i = 0; i < items_array_player.size(); i++) {
-        gr_inventory_player[i]->slot = items_array_player[i];
-        if (gr_inventory_player[i]->slot->get_item() != nullptr) {
-            float x = gr_inventory_player[i]->slot_sprite.getPosition().x + 16.f;
-            float y = gr_inventory_player[i]->slot_sprite.getPosition().y + 16.f;
-            gr_inventory_player[i]->slot->get_item()->get_sprite().setPosition({x, y});
-            gr_inventory_player[i]->slot->get_item()->get_sprite().setScale({1.25f, 1.25f});
-            gr_inventory_player[i]->gr_amount.setString(std::to_string(gr_inventory_player[i]->slot->get_amount()));
-            if (gr_inventory_player[i]->slot->get_amount() >= 10 && !gr_inventory_player[i]->gr_amount_offset) {
-                float x = gr_inventory_player[i]->gr_amount.getPosition().x - 6.f;
-                float y = gr_inventory_player[i]->gr_amount.getPosition().y;
-                gr_inventory_player[i]->gr_amount.setPosition(sf::Vector2f({x, y}));
-                gr_inventory_player[i]->gr_amount_offset = true;
+        (*gr_inventory_player)[i]->slot = items_array_player[i];
+        if ((*gr_inventory_player)[i]->slot->get_item() != nullptr) {
+            float x = (*gr_inventory_player)[i]->slot_sprite.getPosition().x + 16.f;
+            float y = (*gr_inventory_player)[i]->slot_sprite.getPosition().y + 16.f;
+            (*gr_inventory_player)[i]->slot->get_item()->get_sprite().setPosition({x, y});
+            (*gr_inventory_player)[i]->slot->get_item()->get_sprite().setScale({1.25f, 1.25f});
+            (*gr_inventory_player)[i]->gr_amount.setString(
+                    std::to_string((*gr_inventory_player)[i]->slot->get_amount()));
+            if ((*gr_inventory_player)[i]->slot->get_amount() >= 10 && !(*gr_inventory_player)[i]->gr_amount_offset) {
+                float x = (*gr_inventory_player)[i]->gr_amount.getPosition().x - 6.f;
+                float y = (*gr_inventory_player)[i]->gr_amount.getPosition().y;
+                (*gr_inventory_player)[i]->gr_amount.setPosition(sf::Vector2f({x, y}));
+                (*gr_inventory_player)[i]->gr_amount_offset = true;
             }
         }
     }
@@ -95,15 +98,15 @@ void TradeMenu::update_graphic_inventories(const std::vector<std::shared_ptr<Slo
         }
     }
 
-    gr_money_player.setString("$ " + std::to_string(_money_player));
+    gr_money_player->setString("$ " + std::to_string(_money_player));
     gr_money_trader.setString("$ " + std::to_string(_money_trader));
 }
 
 void TradeMenu::show_inventories(sf::RenderWindow& window) {
-    for (auto& el : gr_inventory_player.gr_items_array)
+    for (auto& el : gr_inventory_player->gr_items_array)
         el->show_slot(window);
     for (auto& el : gr_inventory_trader.gr_items_array)
         el->show_slot(window);
-    window.draw(gr_money_player);
+    window.draw(*gr_money_player);
     window.draw(gr_money_trader);
 }
